Refactor app component to pass through methods
