+++ src/App/App.js
@@ -1,26 +1,33 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import React, { Component } from 'react';
+import './App.scss';
+import Habitat from '../Habitat/Habitat';
+import Form from '../Form/Form';
 
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
+export default class App extends Component {
+  constructor() {
+    super();
+    this.state = {
+      critters: []
+    }
+  }
+
+  rescueCritter = critter => {
+    this.setState({critters: [...this.state.critters, critter]});
+  }
+
+  releaseCritter = id => {
+    const remainingCritters = this.state.critters.filter(critter => {
+      return critter.id !== id;
+    })
+    this.setState({ critters: remainingCritters })
+  }
+
+  render() {
+    return (
+      <section className="app">
+        <Form rescueCritter={this.rescueCritter} />
+        <Habitat critters={this.state.critters} releaseCritter={this.releaseCritter} />
+      </section>
+    );
+  }
 }
-
-export default App;
